7.1. Refactor reducer. Split reducer logic
