+++ src/index.js
@@ -3,11 +3,11 @@ import ReactDOM from 'react-dom';
 import {createStore} from 'redux';
 import {Provider} from 'react-redux';
 import App from './components/app/app.jsx';
-import {reducer} from './store/reducer.js';
+import rootReducer from './store/reducers/root-reducer.js';
 
 
 const store = createStore(
-    reducer,
+    rootReducer,
     window.__REDUX_DEVTOOLS_EXTENSION__
       ? window.__REDUX_DEVTOOLS_EXTENSION__()
       : (f) => f
