+++ src/store/action.js
@@ -1,7 +1,12 @@
 const ActionType = {
+  LOAD_FILMS: `LOAD_FILMS`,
   SELECT_GENRE: `SELECT_GENRE`,
 };
 
+const loadFilms = (films) => ({
+  type: ActionType.LOAD_FILMS,
+  payload: films,
+})
 
 const selectGenre = (genre) => ({
   type: ActionType.SELECT_GENRE,
@@ -11,5 +16,6 @@ const selectGenre = (genre) => ({
 
 export {
   ActionType,
+  loadFilms,
   selectGenre,
 };
